web: removed the use of a drawer for linking connectors.
