+++ web/index.html
@@ -24,7 +24,7 @@
               </div>
           </div>
       </noscript>
-    <div vaul-drawer-wrapper id="app"></div>
+    <div id="app"></div>
     <script type="module" src="/src/main.js"></script>
   </body>
 </html>
