+++ web/package.json
@@ -30,7 +30,6 @@
     "radix-vue": "^1.7.3",
     "tailwind-merge": "^2.2.1",
     "tailwindcss-animate": "^1.0.7",
-    "vaul-vue": "^0.1.2",
     "vue": "^3.4.15",
     "vue-matomo": "^4.2.0",
     "vue-router": "^4.2.5",
