--- web/src/components/ui/drawer/DrawerContent.vue
@@ -1,43 +1,0 @@
-<script setup>
-import { DrawerContent, DrawerPortal } from "vaul-vue";
-import { useForwardPropsEmits } from "radix-vue";
-import DrawerOverlay from "./DrawerOverlay.vue";
-import { cn } from "@/lib/utils";
-
-const props = defineProps({
-  forceMount: { type: Boolean, required: false },
-  trapFocus: { type: Boolean, required: false },
-  disableOutsidePointerEvents: { type: Boolean, required: false },
-  asChild: { type: Boolean, required: false },
-  as: { type: null, required: false },
-  class: { type: null, required: false },
-});
-const emits = defineEmits([
-  "escapeKeyDown",
-  "pointerDownOutside",
-  "focusOutside",
-  "interactOutside",
-  "openAutoFocus",
-  "closeAutoFocus",
-]);
-
-const forwarded = useForwardPropsEmits(props, emits);
-</script>
-
-<template>
-  <DrawerPortal>
-    <DrawerOverlay />
-    <DrawerContent
-      v-bind="forwarded"
-      :class="
-        cn(
-          'fixed inset-x-0 bottom-0 z-50 mt-24 flex h-auto flex-col rounded-t-[10px] border bg-background',
-          props.class
-        )
-      "
-    >
-      <div class="mx-auto mt-4 h-2 w-[100px] rounded-full bg-muted" />
-      <slot />
-    </DrawerContent>
-  </DrawerPortal>
-</template>
