--- web/src/components/ui/drawer/DrawerDescription.vue
@@ -1,26 +1,0 @@
-<script setup>
-import { DrawerDescription } from "vaul-vue";
-import { computed } from "vue";
-import { cn } from "@/lib/utils";
-
-const props = defineProps({
-  asChild: { type: Boolean, required: false },
-  as: { type: null, required: false },
-  class: { type: null, required: false },
-});
-
-const delegatedProps = computed(() => {
-  const { class: _, ...delegated } = props;
-
-  return delegated;
-});
-</script>
-
-<template>
-  <DrawerDescription
-    v-bind="delegatedProps"
-    :class="cn('text-sm text-muted-foreground', props.class)"
-  >
-    <slot />
-  </DrawerDescription>
-</template>
