--- web/src/components/ui/drawer/DrawerHeader.vue
@@ -1,13 +1,0 @@
-<script setup>
-import { cn } from "@/lib/utils";
-
-const props = defineProps({
-  class: { type: null, required: false },
-});
-</script>
-
-<template>
-  <div :class="cn('grid gap-1.5 p-4 text-center sm:text-left', props.class)">
-    <slot />
-  </div>
-</template>
