--- web/src/components/ui/drawer/DrawerOverlay.vue
@@ -1,25 +1,0 @@
-<script setup>
-import { DrawerOverlay } from "vaul-vue";
-import { computed } from "vue";
-import { cn } from "@/lib/utils";
-
-const props = defineProps({
-  forceMount: { type: Boolean, required: false },
-  asChild: { type: Boolean, required: false },
-  as: { type: null, required: false },
-  class: { type: null, required: false },
-});
-
-const delegatedProps = computed(() => {
-  const { class: _, ...delegated } = props;
-
-  return delegated;
-});
-</script>
-
-<template>
-  <DrawerOverlay
-    v-bind="delegatedProps"
-    :class="cn('fixed inset-0 z-50 bg-black/80', props.class)"
-  />
-</template>
