--- web/src/components/ui/drawer/index.js
@@ -1,8 +1,0 @@
-export { DrawerPortal, DrawerTrigger, DrawerClose } from "vaul-vue";
-export { default as Drawer } from "./Drawer.vue";
-export { default as DrawerOverlay } from "./DrawerOverlay.vue";
-export { default as DrawerContent } from "./DrawerContent.vue";
-export { default as DrawerHeader } from "./DrawerHeader.vue";
-export { default as DrawerFooter } from "./DrawerFooter.vue";
-export { default as DrawerTitle } from "./DrawerTitle.vue";
-export { default as DrawerDescription } from "./DrawerDescription.vue";
